+++ package.json
@@ -58,10 +58,9 @@
     "react-dom": "^15"
   },
   "dependencies": {
-    "@helpscout/fancy": "^2.0.3",
-    "@helpscout/react-utils": "^0.0.3",
+    "@helpscout/fancy": "^2.0.4",
+    "@helpscout/react-utils": "^1.0.0",
     "closest": "0.0.1",
-    "create-react-context": "0.2.2",
     "emoji-mart": "^1.0.1",
     "popper.js": "1.14.3",
     "prop-types": "15.6.1",
