+++ src/components/PropProvider/Context.js
@@ -1,5 +1,6 @@
 // @flow
-import createContext from 'create-react-context'
+import createContext from '@helpscout/react-utils/dist/createContext'
+
 import { contextConfig } from './utils'
 
 const Context = createContext(contextConfig)
