+++ src/components/PropProvider/README.md
@@ -5,4 +5,4 @@ A Provider component provides child Blue components with config values.
 ## Usage
 
 * [Provider](./docs/Provider.md)
-* [configConnect](./docs/configConnect.md)
+* [propConnect](./docs/propConnect.md)
