+++ src/components/PropProvider/docs/propConnect.md
@@ -1,4 +1,4 @@
-# configConnect
+# propConnect
 
 Higher-order function that "connects" a React component with Blue's config, generated by the [PropProvider](./Provider.md). This pattern is similar to [Redux's connect](https://github.com/reduxjs/react-redux).
 
@@ -6,19 +6,19 @@ Higher-order function that "connects" a React component with Blue's config, gene
 
 ```jsx
 import React, { Component } from 'react'
-import { configConnect } from '@helpscout/blue/PropProvider'
+import { propConnect } from '@helpscout/blue/PropProvider'
 
 const MyComponent extends Component {
   ...
 }
 
-export default configConnect('MyComponent')(MyComponent)
+export default propConnect('MyComponent')(MyComponent)
 ```
 
 ## Arguments
 
 ```
-configConnect(name)(Component)
+propConnect(name)(Component)
 ```
 
 | Argument    | Type              | Description                                              |
