+++ src/components/PropProvider/index.js
@@ -3,6 +3,6 @@ import Provider from './Provider'
 export { default as Provider } from './Provider'
 export { default as Context } from './Context'
 export { default as Consumer } from './Consumer'
-export { default as configConnect } from './configConnect'
+export { default as propConnect } from './propConnect'
 
 export default Provider
